docs/examples/ignore/sorting.py
# type: ignore
# ruff: noqa E402
"""This file demonstrates how to sort a natural language string in a "verbalized algorithm" style.

Verbalized algorithms is a style of algorithms that uses LLM to perform
high-level reasoning tasks such as sorting, clustering, etc.
Given an informal representation of a formal task,
it operates by replacing some basic atomic operations in a classical algorithm
with a LLM-driven text-based one while retaining the high-level control flow,
thereby largely retaining the theoretical guarantees of the original algorithm.

Imagine receiving hundreds of customer complaints and one must quickly
sort and select 5 customers who most urgently need supports.
Feeding such a large set of input strings blindly into an LLM is ineffective and impractical
both due to the limited context length and its limited ability to perform sorting task
--- LLMs provide no guarantees on the correctness of their outputs for a computational reasoning task such as sorting,
other than anecdotal, empirical observations that they sometimes do.

In contrast, traditional disciplines of computer science approach this task
using algorithms that involve task decomposition.
For example, a merge sort decomposes the entire sorting task into several sorting tasks over individual sublists
and combines the results, where at the limit, there is an elementary binary comparison operator that is applied to
individual pair of objects in the list.

"""

import random

import mellea
from mellea.backends import Backend, CBlock, Component, Context, ModelOutputThunk
from mellea.backends.types import ModelOption
from mellea.stdlib.base import LinearContext, SimpleContext
from mellea.stdlib.requirement import Requirement

m = mellea.start_session()

import json
from typing import Literal, Optional, Type, TypeVar  # noqa: UP035

from pydantic import BaseModel

M = TypeVar("M", bound=BaseModel)


def askdata(query: list[str] | str, cls: type[M], n=5, verbose=True) -> M | list[M]:
    """Poor-man's structured decoding"""

    system_message = (
        " You are an assistant that must always return a json string that follows a given json schema."
        + " The schema is this: "
        + json.dumps(cls.model_json_schema())
    )

    if isinstance(query, str):
        queries, need_unbatch = [query], True
    else:
        queries, need_unbatch = query, False

    results = []
    for q in queries:
        res = m.instruct(
            q, model_options={ModelOption.SYSTEM_PROMPT: system_message}, format=cls
        )
        obj = cls.model_validate_json(res.value)  # type:ignore
        results.append(obj)
        if verbose:
            bp(f"{q} -> ")
            bp(obj.model_dump_json(indent=2))

    if need_unbatch:
        return results[0]
    else:
        return results


def askchoice(query: list[str] | str, choices=list[str], verbose=True):
    """Poor-man's constraint decoding"""

    class Choice(BaseModel):
        think: str
        answer: Literal[*choices]  # type: ignore

    answers = askdata(query, Choice, verbose=verbose)
    if isinstance(answers, list):
        return [c.answer for c in answers]
    else:
        return answers.answer


# printer facility

import textwrap
from collections.abc import Callable, Iterable


def wrapped(text: str | Iterable[str], width: int = 100, fn: Callable = print):
    """
    Print a text string in a reasonable screen width (100).
    Swapping fn with a custom function (e.g. yp, rp, ....) is handy.
    """
    if not isinstance(text, str):
        for elem in text:
            for subline in textwrap.wrap(
                str(elem), width=width, initial_indent="* ", subsequent_indent="  "
            ):
                fn(subline)
    else:
        for line in text.split("\n"):
            for subline in textwrap.wrap(line, width=width):
                fn(subline)


from colors import blue, cyan, green, magenta, red, yellow


def yp(line):
    "Print the line in yellow."
    print(yellow(line), flush=True)


def rp(line):
    "Print the line in red."
    print(red(line), flush=True)


def gp(line):
    "Print the line in green."
    print(green(line), flush=True)


def bp(line):
    "Print the line in blue."
    print(blue(line), flush=True)


def mp(line):
    "Print the line in magenta."
    print(magenta(line), flush=True)


def cp(line):
    "Print the line in cyan."
    print(cyan(line), flush=True)


# verbalized sorting.

import abc
from functools import partial


class Parameterized(abc.ABC):
    """Abstract base class for string objects with LLM-defined algebraic structures."""

    def __class_getitem__(cls, key):
        if not isinstance(key, tuple):
            key = (key,)
        return type(f"{cls.__name__}[{key!r}]", (cls,), {"args": key})


class Comparable(Parameterized):
    """Makes strings compatible with operator overloading.

    Specify the ordering criteria by scripting.
    For example, the following code creates a class
    whose instances are compared by the amount of anger in the message.

    Sentiment = Comparable["Is X angrier than Y?"]

    """

    def __init__(self, string: str):
        self.string = string

    @classmethod
    @property
    def criteria(cls):
        return cls.args[0]

    def __gt__(self, other):
        return (
            askchoice(
                f"{self.criteria} X: '{self.string}' Y: '{other.string}'", ["yes", "no"]
            )
            == "yes"
        )

    def __repr__(self):
        return f'{self.__class__.__name__}("{self.string}")'


def main():
    Sentiment = Comparable[
        "On the scale of angry/frustrated versus happy/relaxed, is X angrier and/or less happy than Y?"
    ]
    print(Sentiment)
    Sentiment = Comparable[
        (
            "On the scale of angry/frustrated versus happy/relaxed, is X angrier and/or less happy than Y?",
        )
    ]  # equivalent
    print(Sentiment)
    Sentiment = Comparable[
        "On the scale of angry/frustrated versus happy/relaxed, is X angrier and/or less happy than Y?",
    ]  # equivalent
    print(Sentiment)

    x = Sentiment("The server is down!!! How can this be possible!? We paid you a lot!")
    y = Sentiment(
        "It seems the server is down as usual. Quite unfortunate. But it is what it is."
    )
    z = Sentiment(
        "Yey, the server is down, I can call it a day and go home to enjoy some beer."
    )
    # w = Sentiment(
    #     "When you came, you said to me as follows: 'I will give Gimil-Sin (when he comes) fine quality copper ingots.' You left then but you did not do what you promised me."
    # )

    print(x)

    rp(f"Verbalized sorting by: {Sentiment.criteria}")
    xs = [x, y, z]
    print("original:")
    for elem in xs:
        print(elem.string)
    print("original/sorted:")
    for elem in sorted(xs):
        print(elem.string)
    random.shuffle(xs)
    print("shuffled:")
    for elem in xs:
        print(elem.string)
    print("shuffled/sorted:")
    for elem in sorted(xs):
        print(elem.string)

    Sentiment = Comparable[
        "Does the sentence X indicate more positive emotion toward Bob than Y does?"
    ]
    rp(f"Verbalized sorting by: {Sentiment.criteria}")

    x = Sentiment("I love Bob.")
    y = Sentiment("I like Bob.")
    z = Sentiment("Bob is meh.")
    w = Sentiment("I hate Bob.")

    xs = [x, y, z, w]
    print("original:")
    for elem in xs:
        print(elem.string)
    print("original/sorted:")
    for elem in sorted(xs):
        print(elem.string)
    random.shuffle(xs)
    print("shuffled:")
    for elem in xs:
        print(elem.string)
    print("shuffled/sorted:")
    for elem in sorted(xs):
        print(elem.string)

    Color = Comparable[
        "Is the svg color X brighter than the svg color Y? Consider the HSV color value. "
    ]
    rp(f"Verbalized sorting by: {Color.criteria}")

    xs = list(
        map(
            Color,
            [
                "Red",
                "Blue",
                "Green",
                "Yellow",
                "Orange",
                "Purple",
                "Pink",
                "Brown",
                "Black",
                "White",
                "Gray",
                "Cyan",
                "Magenta",
                "Lime",
                "Navy",
                "Teal",
                "Maroon",
                "Olive",
                "Indigo",
                "Gold",
            ],
        )
    )
    random.shuffle(xs)
    print("shuffled:")
    for elem in xs:
        print(elem.string)
    print("shuffled/sorted:")
    for elem in sorted(xs):
        print(elem.string)

    Number = Comparable["Is X larger than Y as a decimal floating-point number?"]
    rp(f"Verbalized sorting by: {Number.criteria}")

    x = Number("1.2")
    y = Number("9.9")
    z = Number("9.11")
    w = Number("7.3")

    l = [x, y, z, w]
    for elem in sorted(l):
        print(elem.string)

    Version = Comparable[
        "Is X larger than Y as a string following semantic versioning?"
    ]
    rp(f"Verbalized sorting by: {Version.criteria}")

    x = Version("1.2")
    y = Version("9.9")
    z = Version("9.11")
    w = Version("7.3")

    l = [x, y, z, w]
    for elem in sorted(l):
        print(elem.string)


if __name__ == "__main__":
    # import fattrace
    # fattrace.install(include_external=True)

    main()
