pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "mellea"
version = "0.0.1"
authors = [
    { name = "Nathan Fulton", email = "nathan@ibm.com" },
    { name = "Hendrik Strobelt", email = "hendrik.strobelt@ibm.com" },
    { name = "Avinash Balakrishnan", email = "avinash.bala@us.ibm.com" },
    { name = "Elron Bandel", email = "elron.bandel@ibm.com" },
    { name = "Masataro Asai", email = "masataro.asai@ibm.com" },
    { name = "Supriya Lall", email = "supriya.lall@partner.ibm.com" },
    { name = "Xiaojie Guo", email = "xiaojie.guo@ibm.com" },
    { name = "Jake LoRocco", email = "jake.lorocco@ibm.com" }
]
description = "mellea is a librarty for writing generative programs"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pydantic",
    "outlines<1.0.0",
    "openai",
    "jinja2",
    "transformers",
    "torchvision",
    "accelerate>=0.26.0",
    "json5",
    "docling==2.28.1",
    "ollama>=0.5.1",
    "alora==0.2.0",
    "ansicolors",
    "datasets",
    "ibm-watsonx-ai",
    "requests>=2.32.4",
    "uvicorn",
    "fastapi",
    "types-requests",
    "types-tqdm",
    "typer",
    "click<8.2.0", # Newer versions will cause errors with --help in typer CLIs.
    "mistletoe>=1.4.0",
    "trl",
    "peft",
    "torch"
]

[project.scripts]
m = "cli.m:cli"

[project.optional-dependencies]
# Use these like:
# pip install -e .[dev]
# pip install generative_computing[dev]
dev = ["mypy", "mypy-extensions", "tomli", "codespell"]
sphinx = ["sphinx_rtd_theme", "sphinx-autodoc-typehints", "sphinx_mdinclude"]

[dependency-groups]
# Use these like:
# pip install -e . --group dev
# pip install generative_computing --group dev
dev = [
    "isort>=6.0.0",
    "pre-commit>=4.2.0",
    "pylint>=3.3.4",
    "ruff>=0.11.6",
    "pdm>=2.24.0",
    "pytest",
    "mypy>=1.17.0",
]
docling = [
    "docling>=2.15.1",
]
all = [
    "mellea",
    "mellea[dev]"
]
notebook = [
    "ipykernel>=6.29.5",
    "ipython>=8.36.0",
    "jupyter>=1.1.1",
]

sphinx = [
    "sphinx_rtd_theme",
    "sphinx-autodoc-typehints",
    "sphinx_mdinclude",
]

[tool.ruff]
target-version = "py312"
respect-gitignore = true

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = [
  # "B",  # flake8-bugbear
  "C",  # flake8-comprehensions
  "C9", # mccabe
  # "D",  # flake8-docstrings
  "E",   # pycodestyle errors (default)
  "F",   # pyflakes (default)
  "I",   # isort
  "PD",  # pandas-vet
  "PIE", # pie
  # "PTH", # pathlib
  "Q", # flake8-quotes
  # "RET", # return
  "RUF", # Enable all ruff-specific checks
  # "SIM", # simplify
  "S307", # eval
  # "T20",  # (disallow print statements) keep debugging statements out of the codebase
  "W",     # pycodestyle warnings
  "ASYNC", # async
  "UP",    # pyupgrade
]

ignore = [
  "RUF001", # ambiguous unicode characters (these occur in examples).
  "C408",   # Unnecessary `dict()` call (rewrite as a literal)
  "E501",   # Line too long, handled by ruff formatter
  # "D107",   # "Missing docstring in __init__",
  "F401",   # imported but unused; consider using `importlib.util.find_spec` to test for "
  "F811",   # "redefinition of the same function"
  "PL",     # Pylint
  "RUF012", # Mutable Class Attributes
#   "UP006",  # List vs list, etc
#   "UP007",  # Option and Union
#   "UP035",  # `typing.Set` is deprecated, use `set` instead"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = false

[tool.codespell]
ignore-words-list = 'mellea,hashi,noo,Asai,asai'
check-filenames = true
check-hidden = false
regex = "(?<![a-z])[a-z'`]+|[A-Z][a-z'`]*|[a-z]+'[a-z]*|[a-z]+(?=[_-])|[a-z]+(?=[A-Z])|\\d+"
skip = 'requirements.txt,uv.lock'

[tool.mypy]
disable_error_code = ["empty-body", "import-untyped"]
